+++ apps/docs/src/examples/button/loading.tsx
@@ -1,0 +1,10 @@
+import { Button } from "@pixelshades/ui/components"
+import { SnailIcon } from "@pixelshades/ui/icons"
+
+export default function Example() {
+	return (
+		<Button before={<SnailIcon />} isLoading>
+			Button
+		</Button>
+	)
+}