+++ apps/docs/src/examples/button/preview.tsx
@@ -2,5 +2,5 @@ import { Button } from "@pixelshades/ui/components"
 import { SnailIcon } from "@pixelshades/ui/icons"
 
 export default function Example() {
-	return <Button before={<SnailIcon className="bg-red-600" />}>Button</Button>
+	return <Button before={<SnailIcon />}>Button</Button>
 }