+++ packages/styles/src/components/date-picker/variants.ts
@@ -8,9 +8,8 @@ import { dateFieldVariants } from "../date-field/variants"
 export const datePickerVariants = tv({
 	base: "",
 	slots: {
-		input: "pr-8",
 		inputButton:
-			"absolute right-2 flex size-5 appearance-none items-center justify-center rounded-md outline-none ring-primary ring-offset-2 ring-offset-surface data-[hovered]:bg-subtle data-[focus-visible]:ring-2",
+			"flex size-4 appearance-none items-center justify-center rounded-md outline-none ring-primary ring-offset-2 ring-offset-surface data-[hovered]:bg-subtle data-[focus-visible]:ring-2",
 	},
 	extend: dateFieldVariants,
 })