+++ packages/ui/src/components/ui/checkbox-group/checkbox-group.tsx
@@ -7,14 +7,18 @@
 import { checkboxGroup } from "@pixelshades/styles/components/checkbox-group"
 import type { ReactNode } from "react"
 import type { CheckboxGroupProps as AriaCheckboxGroupProps } from "react-aria-components"
-import { CheckboxGroup as AriaCheckboxGroup, Text } from "react-aria-components"
+import { CheckboxGroup as AriaCheckboxGroup } from "react-aria-components"
 import { FormDescription, FormFieldError } from "../form"
 import { type FormComponentLabelProps, Label } from "../label"
 
 interface CheckboxGroupProps extends Omit<AriaCheckboxGroupProps, "className" | "children">, FormComponentLabelProps {
+	/** The className of the checkbox group. */
 	className?: string
+	/** Checkbox or any other components to render in the checkbox group. */
 	children?: ReactNode
+	/** A subtle description next to the checkbox groups label. */
 	helperText?: string
+	/** The error message of the checkbox group. */
 	errorMessage?: string
 }
 
@@ -29,12 +33,16 @@ const CheckboxGroup = ({
 	...props
 }: CheckboxGroupProps) => (
 	<AriaCheckboxGroup {...props} className={checkboxGroup({ className })}>
-		<Label description={description} tooltip={tooltip}>
-			{label}
-		</Label>
-		{children}
-		{description && <FormDescription>{helperText}</FormDescription>}
-		<FormFieldError>{errorMessage}</FormFieldError>
+		{({ isRequired, isDisabled }) => (
+			<>
+				<Label description={description} aria-disabled={isDisabled} tooltip={tooltip} isRequired={isRequired}>
+					{label}
+				</Label>
+				{children}
+				{helperText && <FormDescription aria-disabled={isDisabled}>{helperText}</FormDescription>}
+				<FormFieldError>{errorMessage}</FormFieldError>
+			</>
+		)}
 	</AriaCheckboxGroup>
 )