+++ packages/ui/src/components/ui/date-field/date-field.tsx
@@ -5,15 +5,19 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import type { DateFieldProps as AriaDateFieldProps, DateValue, ValidationResult } from "react-aria-components"
-import { DateField as AriaDateField, DateInput, DateSegment, Text } from "react-aria-components"
+import { DateField as AriaDateField, DateInput, DateSegment } from "react-aria-components"
 
 import { dateFieldVariants } from "@pixelshades/styles/components/date-field"
 import { inputVariants } from "@pixelshades/styles/components/input"
+import { cn } from "@pixelshades/utils/styles"
+import type { ReactNode } from "react"
 import { FormDescription, FormFieldError } from "../form"
 import { type FormComponentLabelProps, Label } from "../label"
 
 interface DateFieldProps<T extends DateValue> extends AriaDateFieldProps<T>, FormComponentLabelProps {
-	helperText?: string
+	/** A helper text to display below the date field. */
+	helperText?: ReactNode
+	/** The error message to display if the date field is invalid. */
 	errorMessage?: string | ((validation: ValidationResult) => string)
 }
 
@@ -23,13 +27,15 @@ const DateField = <T extends DateValue>({
 	tooltip,
 	description,
 	errorMessage,
+	className,
+	isRequired,
 	...props
 }: DateFieldProps<T>) => (
-	<AriaDateField className="flex flex-col gap-md" {...props}>
-		<Label description={description} tooltip={tooltip}>
+	<AriaDateField className={cn("flex flex-col gap-md", className)} isRequired={isRequired} {...props}>
+		<Label description={description} isRequired={isRequired} tooltip={tooltip}>
 			{label}
 		</Label>
-		<DateInput className={inputVariants()}>
+		<DateInput className={inputVariants(props as any)}>
 			{(segment) => <DateSegment className={dateFieldVariants()} segment={segment} />}
 		</DateInput>
 		{helperText && <FormDescription>{helperText}</FormDescription>}