+++ packages/ui/src/components/ui/text-field/text-field.tsx
@@ -4,28 +4,48 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-import { textFieldVariants } from "@pixelshades/styles/components/text-field"
+import { forwardRef } from "@pixelshades/utils/jsx"
+import type { ReactNode } from "react"
 import type { TextFieldProps as AriaTextFieldProps } from "react-aria-components"
 import { TextField as AriaTextField } from "react-aria-components"
 import { FormDescription, FormFieldError } from "../form"
 import { Input } from "../input"
 import { type FormComponentLabelProps, Label } from "../label"
+import { TextArea } from "./text-area"
 
 export interface TextFieldProps extends AriaTextFieldProps, FormComponentLabelProps {
-	label?: string
-	helperText?: string
+	helperText?: ReactNode
 	errorMessage?: string
+
+	/** Text Area Props */
+	multiLine?: boolean
+	cols?: number
+	rows?: number
 }
 
-const TextField = ({ label, helperText, description, tooltip, errorMessage, ...props }: TextFieldProps) => (
-	<AriaTextField className={"flex flex-col gap-sm"} {...props}>
-		<Label tooltip={tooltip} description={description}>
-			{label}
-		</Label>
-		<Input />
-		{helperText && !errorMessage && <FormDescription>{helperText}</FormDescription>}
-		<FormFieldError>{errorMessage}</FormFieldError>
-	</AriaTextField>
+const TextField = forwardRef(
+	({
+		label,
+		helperText,
+		description,
+		tooltip,
+		errorMessage,
+		ref,
+		isRequired,
+		multiLine,
+		cols,
+		rows,
+		...props
+	}: TextFieldProps & { ref?: any }) => (
+		<AriaTextField className={"flex flex-col gap-sm"} isRequired={isRequired} {...props}>
+			<Label tooltip={tooltip} description={description} isRequired={isRequired}>
+				{label}
+			</Label>
+			{multiLine ? <TextArea ref={ref} cols={cols} rows={rows} /> : <Input ref={ref} />}
+			{helperText && <FormDescription>{helperText}</FormDescription>}
+			<FormFieldError>{errorMessage}</FormFieldError>
+		</AriaTextField>
+	),
 )
 
 export { TextField }