+++ apps/docs/CHANGELOG.md
@@ -1,5 +1,21 @@
 # docs
 
+## 0.0.3
+
+### Patch Changes
+
+- Updated dependencies [fbb7db0]
+  - @pixelshades/styles@0.0.3
+  - @pixelshades/ui@0.0.7
+
+## 0.0.2
+
+### Patch Changes
+
+- Updated dependencies [14dfe06]
+  - @pixelshades/styles@0.0.1
+  - @pixelshades/ui@0.0.5
+
 ## 0.0.1
 
 ### Patch Changes